+++ src/h264grabber/.gitignore
@@ -1,0 +1,3 @@
+# Ignore the install dir
+_install/
+
