+++ src/h264grabber/cleanup.h264grabber
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+SCRIPT_DIR=$(cd `dirname $0` && pwd)
+cd $SCRIPT_DIR
+
+cd h264grabber
+
+make clean
