+++ src/rRTSPServer/Makefile.rRTSPServer
@@ -69,9 +69,7 @@ livemedia:
 .$(CPP).$(OBJ):
 	$(CPLUSPLUS_COMPILER) -c $(CPLUSPLUS_FLAGS) -o $@ $<
 
-rRTSPServer_OBJS	= src/rRTSPServer.$(OBJ) \
-				src/ByteStreamCBMemorySource.$(OBJ) src/H264VideoCBMemoryServerMediaSubsession.$(OBJ) \
-				src/CBMemoryServerMediaSubsession.$(OBJ)
+rRTSPServer_OBJS	= src/rRTSPServer.$(OBJ)
 
 rRTSPServer$(EXE):	$(rRTSPServer_OBJS) $(LOCAL_LIBS)
 	$(LINK)$@ $(CONSOLE_LINK_OPTS) $(rRTSPServer_OBJS) $(LIBS) -lpthread
