+++ src/rRTSPServer/init.rRTSPServer
@@ -19,4 +19,3 @@ cd live || exit 1
 ./genMakefiles linux-cross
 cp -f ../Makefile.rRTSPServer Makefile
 cp -rf ../src .
-cp -rf ../include .
