+++ src/static/static/home/yi-hack/etc/system.conf
@@ -16,6 +16,7 @@ NTPD=yes
 NTP_SERVER=pool.ntp.org
 PROXYCHAINSNG=no
 RTSP_PORT=554
+RTSP_SUB_PORT=8554
 ONVIF_PORT=80
 HTTPD_PORT=8080
 USERNAME=
