+++ src/www/httpd/htdocs/js/modules/configurations.js
@@ -25,7 +25,7 @@ APP.configurations = (function ($) {
                 loadingStatusElem.fadeOut(500);
 
                 $.each(response, function (key, state) {
-                    if(key=="HOSTNAME" || key=="TIMEZONE" || key=="NTP_SERVER" || key=="HTTPD_PORT" || key=="RTSP_PORT" || key=="ONVIF_PORT" || key=="USERNAME")
+                    if(key=="HOSTNAME" || key=="TIMEZONE" || key=="NTP_SERVER" || key=="HTTPD_PORT" || key=="RTSP_PORT" || key=="RTSP_SUB_PORT" || key=="ONVIF_PORT" || key=="USERNAME")
                         $('input[type="text"][data-key="' + key +'"]').prop('value', state);
                     else if(key=="RTSP_STREAM" || key=="RTSP_AUDIO" || key=="ONVIF_PROFILE")
                         $('select[data-key="' + key +'"]').prop('value', state);
@@ -68,6 +68,7 @@ APP.configurations = (function ($) {
         configs["RTSP_STREAM"] = $('select[data-key="RTSP_STREAM"]').prop('value');
         configs["RTSP_AUDIO"] = $('select[data-key="RTSP_AUDIO"]').prop('value');
         configs["RTSP_PORT"] = $('input[type="text"][data-key="RTSP_PORT"]').prop('value');
+        configs["RTSP_SUB_PORT"] = $('input[type="text"][data-key="RTSP_SUB_PORT"]').prop('value');
         configs["ONVIF_PORT"] = $('input[type="text"][data-key="ONVIF_PORT"]').prop('value');
         configs["ONVIF_PROFILE"] = $('select[data-key="ONVIF_PROFILE"]').prop('value');
         configs["USERNAME"] = $('input[type="text"][data-key="USERNAME"]').prop('value');
