+++ src/www/httpd/htdocs/pages/configurations.html
@@ -285,6 +285,15 @@
                         </td>
                     </tr>
                     <tr class="row">
+                        <td>RTSP Sub Port</td>
+                        <td>
+                            <input class="u-full-width" type="text" placeholder="" data-key="RTSP_SUB_PORT" id="RTSP1_PORT"/>
+                            <span class="switch-description">
+                                TCP port used by RTSP Server sub stream (default 8554).
+                            </span>
+                        </td>
+                    </tr>
+                    <tr class="row">
                         <td>ONVIF Port</td>
                         <td>
                             <input class="u-full-width" type="text" placeholder="" data-key="ONVIF_PORT" id="ONVIF_PORT"/>
